examples/websockets_lib/requirements.txt
graphql_ws
sanic>=0.7.0
graphene>=2.0
